split query classes into multiple files
